refactor(llm): share retry-after parsing
